make conversion explicit for inevitable future problems

+++ command_handler.cpp
@@ -48,7 +48,9 @@ void managed_duktape_thread(unsafe_info* info)
 {
     int id = get_thread_id(info->ctx);
 
-    mongo_lock_proxy::thread_id_storage_hack = id;
+    ///set thread storage hack
+    ///convert from int to size_t
+    mongo_lock_proxy::thread_id_storage_hack = (size_t)id;
 
     if(duk_safe_call(info->ctx, unsafe_wrapper, (void*)info, 0, 1) != 0)
     {
